requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
dj-database-url==1.0.0
dj-email-url==1.0.6
Django==4.1.1
django-allauth==0.50.0
django-cache-url==3.4.2
django-cors-headers==3.14.0
django-debug-toolbar==3.4.0
django-pure-pagination==0.3.0
djangorestframework==3.14.0
environs==9.5.0
gunicorn==20.1.0
idna==3.4
marshmallow==3.18.0
oauthlib==3.2.1
packaging==21.3
Pillow==9.0.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
types-cryptography==3.3.23
urllib3==1.26.12
django-tinymce==3.5.0
boto3==1.24.22
django-storages==1.12